feat: 오목 게임 진행 중 상태 추가
